Tutorial ode collision 7: convert number of particles to int
Removes the following warnings:
ode-collision-7-hit_func-3-contacts.py:204: DeprecationWarning: integer argument expected, got float
  s = soya.Smoke(self.parent,nb_particles=12*nl)
ode-collision-7-hit_func-3-contacts.py:211: DeprecationWarning: integer argument expected, got float
  s = soya.Smoke(self.parent,nb_particles=12*nl)

+++ soya/tutorial/ode-collision-7-hit_func-3-contacts.py
@@ -198,7 +198,7 @@ class LaserCamera(soya.Camera):
 				target.add_force(self.laser_vector*self.laser_power,impact)
 				
 				l = sqrt(self.laser_power/1000.)
-				nl = sqrt(l)
+				nl = int(sqrt(l))
 				#create particle
 				# first corresponding to the dust of the planet
 				s = soya.Smoke(self.parent,nb_particles=12*nl)
